♻️  Refactor Custom Table Cells
