+++ components/TableCells/VariantsFileIdCell.tsx
@@ -1,0 +1,39 @@
+/*
+ *
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ *  This program and the accompanying materials are made available under the terms of
+ *  the GNU Affero General Public License v3.0. You should have received a copy of the
+ *  GNU Affero General Public License along with this program.
+ *   If not, see <http://www.gnu.org/licenses/>.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ *  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ *  SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ *  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ *  OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ *  IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ */
+
+import { get } from 'lodash';
+import urlJoin from 'url-join';
+
+import { FILES_PATH } from '../../global/utils/constants';
+import StyledLink from '../Link';
+import { getConfig } from '../../global/config';
+import { sqonUrlQueryString } from '../../global/utils/sqon';
+
+const VariantsFileIdCell = (props: any) => {
+  const { NEXT_PUBLIC_BASE_PATH } = getConfig();
+  const value = get(props, "original.occurrence.hits.edges[0].node.file.object_id");
+  const fileLink = urlJoin(NEXT_PUBLIC_BASE_PATH, `${FILES_PATH}/?${sqonUrlQueryString("object_id", value)}`);
+  return value ? (
+    <StyledLink href={fileLink}>{value}</StyledLink>
+  ) : null;
+};
+
+export default VariantsFileIdCell;
