+++ components/TableCells/index.tsx
@@ -1,0 +1,2 @@
+export { default as FilesObjectIdCell } from './FilesObjectIdCell';
+export { default as VariantsFileIdCell } from './VariantsFileIdCell';
