+++ global/utils/sqon.ts
@@ -1,0 +1,3 @@
+export const sqonUrlQueryString = (field: string, value: string)=> (
+  `sqon={"op":"and","content":[{"op":"in","content":{"field":"${field}","value":["${value}"]}}]}`
+);
